+++ src/config.ts
@@ -32,6 +32,13 @@ class SdkConfig extends EventEmitter {
   session = true;
 
   /**
+   * Enables client-side caching of requests for tiles and fonts.
+   * The cached requests persist multiple browser sessions and will be reused when possible.
+   * Works only for requests to the MapTiler Cloud API when sessions are enabled.
+   */
+  caching = true;
+
+  /**
    * Unit to be used
    */
   private _unit: Unit = "metric";