+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## 2.0.0
 ### New Features
 - Updating with MapLibre GL JS v4.1
+- Added client-side caching mechanism for tiles and fonts
 ### Bug Fixes
 ### Others
 - Removed `Map.loadImageAsync()` as MapLibre's `.loadImage()` is now promise based.
@@ -143,7 +144,7 @@
 ### New Features
 - The `apiKey` can now be specified in the `Map` constructor (will propagate to `config`)
 - The `language` can now be speficifed in the `Map` constructo (will **not** propagete to `config` and will apply only to this specific instance)
-- `Map` now has the method `.getSdkConfig()` to retrieve the config object. 
+- `Map` now has the method `.getSdkConfig()` to retrieve the config object.
 - `Map` now has the method `.getMaptilerSessionId()` to retrieve the MapTiler session ID
 Both `.getSdkConfig()` and `.getMaptilerSessionId()` are handy for layers or control built outside of the SDK that still need some of the configuration to interact with the server. Those components do not always have access to the internal of the SDK (especially that the config is scoped) but can access to the `Map` instance to which they are added with the implementation of the `.onAdd()` method.